Module Mcset

An imperative, mutable, implementation of a totally ordered Set datastructure, with the following properties:

Thread-safe: concurrent updates from multiple domains are possible. These sets can typically be used to synchronize multiple workers on a common goal.

Lock-free: concurrent modifications are able to make progress without waiting for others to finish.

Linearizable: concurrent operations appear to take place sequentially on a linearized timeline, providing a coherent set of elements at all time. It is furthermore possible to take an O(1) snapshot/copy of the set to observe its collection of elements at a given point in time.

module Make (Ord : Stdlib.Set.OrderedType) : sig ... end

Functor building an implementation of the mutable set structure given a totally ordered type.