Module Mcset.Make

Functor building an implementation of the mutable set structure given a totally ordered type.

Parameters

module Ord : Stdlib.Set.OrderedType

Signature

type elt = Ord.t

The type of set elements.

type t

The type of mutable sets.

val empty : unit -> t

empty () returns a new empty set. O(1)

val singleton : elt -> t

singleton x returns a new set containing only the element x. O(1)

val mem : elt -> t -> bool

mem x t returns true if the element x belongs to the set t, false otherwise. O(logN)

val add : elt -> t -> unit

add x t inserts the element x into the set t. If x was already a member, then the set is unchanged. O(logN)

val remove : elt -> t -> bool

remove x t removes the element x from the set t. It returns a boolean indicating if the removal was successful. If false, the element x was already not a member and the set is unchanged. O(logN)

Snapshots

Concurrent modifications of a set are linearizable. The snapshot/copy provides a coherent view of the elements of a set along this linearized timeline.

Further updates to the original set (or its copies) will trigger a minimal copy-on-write of the internal substructures of the set. This doesn't impact the time complexity of any operations, but induces a corresponding memory complexity for copying the modified subparts once.

val copy : t -> t

copy t returns an independently mutable copy of the set t. Further modifications of the set t will not affect its copies (and vice-versa.) O(1)

module View : sig ... end

A read-only, non-mutable view of a set (similar to the purely functional Stdlib.Set.)

val snapshot : t -> View.t

snapshot t returns a read-only view of the elements of the set t. O(1)

val to_view : t -> View.t

Same as snapshot.

val of_view : View.t -> t

of_view v returns a new mutable set containing all the elements of the view v. O(1)

Iterators

The following functions all proceed on a coherent snapshot of their set t argument, created at the start of their execution. Concurrent modifications of the original set t during their execution will not be observed by these traversals.

val cardinal : t -> int

cardinal t returns the number of unique elements in the set t. O(N)

val fold : ( elt -> 'a -> 'a ) -> t -> 'a -> 'a

fold f t init computes f x0 init |> f x1 |> ... |> f xN, where x0 ... xN are the elements of the set t in increasing order. Returns init if the set t was empty.

val iter : ( elt -> unit ) -> t -> unit

iter f t calls f on all the elements of the set t in increasing order.

val for_all : ( elt -> bool ) -> t -> bool

for_all predicate t returns true when all the elements of the set t satisfies the predicate. No ordering is guaranteed and the function will exit early if it finds an invalid element.

val exists : ( elt -> bool ) -> t -> bool

exists predicate t returns true when at least one element of the set t satisfies the predicate. No ordering is guaranteed and the function will exit early if it finds a valid element.

val elements : t -> elt list

elements t is the list of all the elements of t in increasing order.

val to_list : t -> elt list

Same as elements.

val of_list : elt list -> t

of_list lst is the set containing all the elements of the list lst.

val to_seq : t -> elt Stdlib.Seq.t

to_seq t is the sequence containing all the elements of the set t in increasing order. O(1) creation then O(1) amortized for each consumed element of the sequence (with O(logN) worst-case)

val to_rev_seq : t -> elt Stdlib.Seq.t

to_rev_seq t is the sequence containing all the elements of the set t in decreasing order.

val to_seq_from : elt -> t -> elt Stdlib.Seq.t

to_seq_from x t is the sequence containing all the elements of the set t that are larger than or equal to x, in increasing order.

val of_seq : elt Stdlib.Seq.t -> t

of_set seq is the set containing all the elements of the sequence seq.