Module Make.View

A read-only, non-mutable view of a set (with more operations, similar to the purely functional Stdlib.Set interface)

type elt = Ord.t

The type of set elements.

type t

The type of read-only sets.

val empty : t

The empty set.

val singleton : elt -> t

singleton x returns a set containing only the element x. O(1)

val add : elt -> t -> t

add x t returns a set containing the element x and all the elements of t. If x was already in the set t, then the result is physically equal to t. O(logN)

val remove : elt -> t -> t

remove x t returns a set containing the elements of t without x. If x was not a member of the set t, then the result is physically equal to t. O(logN)

val union : t -> t -> t

union t1 t2 returns a set containing all the elements of t1 and t2. O(N) worst-case

val inter : t -> t -> t

inter t1 t2 returns a set containing the shared elements of t1 and t2. O(N) worst-case

val diff : t -> t -> t

diff t1 t2 returns a set containing the elements of t1 that are not members of the set t2. O(N) worst-case

val map : ( elt -> elt ) -> t -> t

map f t returns a set containing the elements f x0, f x1, ..., f xN where x0 ... xN are all the elements of the set t.

The elements are passed to f in increasing order.

The result is physically equal to t if f always returned a physically equal element. O(NlogN) worst-case

val filter : ( elt -> bool ) -> t -> t

filter predicate t returns the subset of elements of the set t that satistifies the predicate (called in increasing order). The resulting set is physical equal to t if no element was rejected. O(N)

val filter_map : ( elt -> elt option ) -> t -> t

filter_map predicate t returns a set containing the Some elements of f x0, f x1, ..., f xN where x0 ... xN are all the elements of the set t.

The elements are passed to f in increasing order.

The result is physically equal to t if f always returned Some physically equal element. O(NlogN) worst-case

val partition : ( elt -> bool ) -> t -> t * t

partiton predicate t returns two sets, the first one containing all the elements of the set t that satisfies predicate, while the second contains all the rejected ones.

The elements are passed to f in increasing order.

The first set is physically equal to t if f always returned true

The second set is physically equal to t if f always returned false. O(N)

val split : elt -> t -> t * bool * t

split x t returns a triple (smaller, found, larger) such that:

smaller is the subset of elements strictly smaller than x

larger is the subset of elements strictly larger than x

found is true if x is a member of the set t, false otherwise. O(logN)

val pop_min : t -> elt * t

pop_min t returns the smallest element and the other elements of the set t, or raises Not_found if the set t is empty. O(logN)

val pop_min_opt : t -> (elt * t) option

pop_min_opt t returns the smallest element and the other elements of the set t, or None if the set t is empty. O(logN)

val pop_max : t -> elt * t

pop_max t returns the largest element and the other elements of the set t, or raises Not_found if the set t is empty. O(logN)

val pop_max_opt : t -> (elt * t) option

pop_max_opt t returns the largest element and the other elements of the set t, or None if the set t is empty.

Comparisons

val equal : t -> t -> bool

equal t1 t2 returns true if the sets t1 and t2 contain the same elements.

val disjoint : t -> t -> bool

disjoint t1 t2 returns true if the sets t1 and t2 have no elements in common.

val compare : t -> t -> int

compare t1 t2 is a total order, suitable for building sets of sets.

Queries

val is_empty : t -> bool

is_empty t returns true when the set t contains no elements, false if it has at least one member. O(1)

val mem : elt -> t -> bool

mem x t returns true if the element x belongs to the set t, false otherwise. O(logN)

val min_elt : t -> elt

min_elt t returns the smallest element of the set t, or raises Not_found if the set is empty. O(logN)

val min_elt_opt : t -> elt option

min_elt_opt t returns the smallest element of the set t, or None if the set is empty. O(logN)

val max_elt : t -> elt

max_elt t returns the largest element of the set t, or raises Not_found if the set is empty. O(logN)

val max_elt_opt : t -> elt option

max_elt_opt t returns the largest element of the set t, or None if the set is empty. O(logN)

val choose : t -> elt

choose t returns an arbitrary element of the set t, or raises Not_found if the set is empty. O(1)

val choose_opt : t -> elt option

choose_opt t returns an arbitrary element of the set t, or None if the set is empty. O(1)

val find : elt -> t -> elt

find x t returns the element of the set that compares equal to x, or raises Not_found if no such element exists. O(logN)

val find_opt : elt -> t -> elt option

find_opt x t returns the element of the set that compares equal to x, or None if no such element exists. O(logN)

val find_first : ( elt -> bool ) -> t -> elt

find_first predicate t returns the smallest element of the set t that satisfies that monotonically increasing predicate, or raises Not_found if no such element exists. O(logN)

val find_first_opt : ( elt -> bool ) -> t -> elt option

find_first_opt predicate t returns the smallest element of the set t that satisfies that monotonically increasing predicate, or None if no such element exists. O(logN)

val find_last : ( elt -> bool ) -> t -> elt

find_last predicate t returns the largest element of the set t that satisfies that monotonically decreasing predicate, or raises Not_found if no such element exists. O(logN)

val find_last_opt : ( elt -> bool ) -> t -> elt option

find_last_opt predicate t returns the smallest element of the set t that satisfies that monotonically increasing predicate, or None if no such element exists. O(logN)

Iterators

val cardinal : t -> int

cardinal t returns the number of unique elements in the set t. O(N)

val fold : ( elt -> 'a -> 'a ) -> t -> 'a -> 'a

fold f t init computes f x0 init |> f x1 |> ... |> f xN, where x0 ... xN are the elements of the set t in increasing order. Returns init if the set t was empty.

val iter : ( elt -> unit ) -> t -> unit

iter f t calls f on all the elements of the set t in increasing order.

val for_all : ( elt -> bool ) -> t -> bool

for_all predicate t returns true when all the elements of the set t satisfies the predicate. No ordering is guaranteed and the function will exit early if it finds an invalid element.

val exists : ( elt -> bool ) -> t -> bool

exists predicate t returns true when at least one element of the set t satisfies the predicate. No ordering is guaranteed and the function will exit early if it finds a valid element.

val elements : t -> elt list

elements t is the list of all the elements of t in increasing order.

val to_list : t -> elt list

Same as elements.

val of_list : elt list -> t

of_list lst is the set containing all the elements of the list lst.

val to_seq : t -> elt Stdlib.Seq.t

to_seq t is the sequence containing all the elements of the set t in increasing order. O(1) creation then O(1) amortized for each consumed element of the sequence (with O(logN) worst-case)

val to_rev_seq : t -> elt Stdlib.Seq.t

to_rev_seq t is the sequence containing all the elements of the set t in decreasing order.

val to_seq_from : elt -> t -> elt Stdlib.Seq.t

to_seq_from x t is the sequence containing all the elements of the set t that are larger than or equal to x, in increasing order.

val of_seq : elt Stdlib.Seq.t -> t

of_set seq is the set containing all the elements of the sequence seq.